+++ app/routes.php
@@ -7,6 +7,8 @@ $router->get('contact', 'PagesController@contact');
 $router->get('features', 'FeaturesController@index');
 $router->get('features/{id}', 'FeaturesController@show');
 $router->post('features', 'FeaturesController@store');
+$router->get('features/{id}/edit', 'FeaturesController@edit');
+$router->post('features/{id}', 'FeaturesController@update');
 $router->post('features/{id}/delete', 'FeaturesController@destroy');
 
 $router->get('admin', 'AdminController@dashboard');