+++ app/views/features/show.view.php
@@ -1,8 +1,8 @@
 <?php require __DIR__.'/../partials/head.php'; ?>
-  <h1 class="margin-bottom-2"><?= $feature->title; ?></h1>
+  <h1 class="text-center margin-bottom-2"><?= $feature->title; ?></h1>
   <div class="grid-x">
     <div class="cell small-10 small-offset-1 medium-8 medium-offset-2">
-      <p class="lead"><?= $feature->description; ?></p>
+      <p class="text-center lead"><?= $feature->description; ?></p>
     </div>
   </div>
 <?php require __DIR__.'/../partials/footer.php'; ?>