+++ app/views/pages/contact.view.php
@@ -1,4 +1,4 @@
 <?php require __DIR__.'/../partials/head.php'; ?>
-  <h1>Contact</h1>
-  <p class="lead">What about a nice form here?</p>
+  <h1 class="text-center">Contact</h1>
+  <p class="lead text-center">What about a nice form here?</p>
 <?php require __DIR__.'/../partials/footer.php'; ?>