+++ app/views/pages/home.view.php
@@ -1,4 +1,4 @@
 <?php require __DIR__.'/../partials/head.php'; ?>
-  <h1>Homepage</h1>
-  <p class="lead">Let’s start a project!</p>
+  <h1 class="text-center">Homepage</h1>
+  <p class="lead text-center">Let’s start a project!</p>
 <?php require __DIR__.'/../partials/footer.php'; ?>